\name{frair-package}
\alias{frair-package}
\alias{frair}
\docType{package}
\title{Functional Response Analysis in R}
\description{Tools to support sensible statistics for functional response analysis.}
\details{
The main workhorses are \code{\link{frair_test}}, \code{\link{frair_fit}}, \code{\link{frair_compare}} and \code{\link{frair_boot}}. These should be the starting points for most users.

}
\author{
Development hosted on GitHub: \url{https://github.com/dpritchard/frair}  

Please file bug reports at: \url{https://github.com/dpritchard/frair/issues}  
}