\name{friar-package}
\alias{friar-package}
\alias{friar}
\docType{package}
\title{Functional Response Analysis in R}
\description{A (draft) package for functional response analysis.}
\details{
The main workhorse is \code{\link{friar_fit}} and this shoudl be the starting point for most users.

\tabular{ll}{
Package: \tab friar\cr
Type: \tab Package\cr
Version: \tab 0.1\cr
Date: \tab 2013-05-18\cr
License: \tab GPL-2\cr
}

}
\author{
Code contributed by: Daniel Pritchard

Package maintained by Daniel Pritchard (www.pritchard.co) 

Hosted on GitHub (https://github.com/dpritchard/friar)
}