\name{frair-package}
\alias{frair-package}
\alias{frair}
\docType{package}
\title{Functional Response Analysis in R}
\description{A (draft) package for functional response analysis.}
\details{
The main workhorse is \code{\link{frair_fit}} and this should be the starting point for most users.

A full vignette with examples will be released with v 0.5, in the mean time users will find help on the relevant help pages.  

}
\author{
Code contributed by: Daniel Pritchard

Package maintained by Daniel Pritchard (www.pritchard.co) 

Hosted on GitHub (https://github.com/dpritchard/frair)

Please file bug reports at: https://github.com/dpritchard/frair/issues
}