\name{frair-package}
\alias{frair-package}
\alias{frair}
\docType{package}
\title{Functional Response Analysis in R}
\description{A (draft) package for functional response analysis.}
\details{
The main workhorses are \code{\link{frair_test}}, \code{\link{frair_fit}}, \code{\link{frair_compare}} and \code{\link{frair_boot}}. These should be the starting points for most users.

A full vignette with examples will be released with version 0.5, in the meantime users will find help on the relevant help pages.  

}
\author{
Code contributed by: Daniel Pritchard  

Package maintained by Daniel Pritchard (www.pritchard.co) 

Development verison hosted on GitHub: \url{https://github.com/dpritchard/frair}  

Please file bug reports at: \url{https://github.com/dpritchard/frair/issues}  
}